Update chains.py:
Added "no_tool_behavior" property: defines how to respond when no tools are detected. Two possible values: (i) "forward": Pass it as a normal message and let Groq return the response as if it were a typical messaging; (ii) "return": It continues as it is solely a request to detect possible functions.